requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.13
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==6.2.1
h11==0.16.0
idna==3.7
oauthlib==3.3.1
playwright==1.53.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.2.2
uvicorn==0.34.3
